requirements.txt
PyQt5
python-vlc 